pyproject.toml
[tool.poetry]
name = "takehome-api"
version = "0.1.0"
description = ""
authors = ["prnvkbr <pranavkuber@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.1.0"
flask = "^3.1.1"
flask-sqlalchemy = "^3.1.1"
flask-cors = "^5.0.1"

[tool.poetry.scripts]
server = "scripts:server"

[tool.poetry.group.dev.dependencies]
sqlalchemy-stubs = "^0.4"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
